refactor: tidy code in jsx2mp-loader
