+++ packages/jsx2mp-loader/src/page-loader.js
@@ -2,7 +2,7 @@ const { readJSONSync, writeJSONSync, writeFileSync, readFileSync, existsSync, mk
 const { relative, join, dirname } = require('path');
 const { getOptions } = require('loader-utils');
 const compiler = require('jsx-compiler');
-const { removeExt } = require('./utils');
+const { removeExt } = require('./utils/pathHelper');
 
 const ComponentLoader = require.resolve('./component-loader');
